@@ -1,6 +1,6 @@
 {
   "name": "@blick.dev/cli",
-  "version": "0.2.0",
+  "version": "0.5.0",
   "description": "cli for blick.dev multi-window responsive design tool for web developers",
   "main": "dist/index.js",
   "scripts": {
@@ -13,17 +13,20 @@
   "author": "blick.dev",
   "license": "MIT",
   "dependencies": {
+    "@blick.dev/app": "0.5.0",
     "@garygrossgarten/cli": "~0.1.0",
+    "electron": "~7.1.3",
     "open": "~7.0.0"
   },
   "devDependencies": {
     "@types/jest": "~24.0.18",
     "@types/minimist": "~1.2.0",
     "@types/node": "~12.7.3",
+    "@types/open": "~6.2.1",
     "jest": "~24.9.0",
+    "prettier": "~1.19.1",
     "ts-jest": "~24.1.0",
-    "typescript": "~3.6.3",
-    "prettier": "~1.19.1"
+    "typescript": "~3.7.3"
   },
   "publisher": "blick.dev",
   "files": [
